Try to change everything #1

+++ domain/src/main/java/com/hellguy39/domain/usecases/eq_settings/SaveEqPresetUseCase.kt
@@ -3,7 +3,7 @@ package com.hellguy39.domain.usecases.eq_settings
 import com.hellguy39.domain.repositories.EqualizerSettingsRepository
 
 class SaveEqPresetUseCase(private val equalizerSettingsRepository: EqualizerSettingsRepository) {
-    operator fun invoke(preset: Int) {
+    operator fun invoke(preset: Short) {
         equalizerSettingsRepository.savePreset(preset = preset)
     }
 }
